requirements.txt
# Core
Django>=5.1,<5.2
django-environ>=0.11,<1.0
psycopg[binary]>=3.1,<4.0
gunicorn>=22.0,<23.0

# Auth
django-allauth[socialaccount]>=65.0,<66.0
bcrypt>=4.2,<5.0

# Background tasks
django-background-tasks>=1.2,<2.0

# Frontend
django-tailwind[reload]>=3.8,<4.0
django-htmx>=1.19,<2.0

# Storage
django-storages[s3]>=1.14,<2.0
Pillow>=10.4,<11.0

# Security
django-csp>=3.8,<4.0
django-ratelimit>=4.1,<5.0
cryptography>=43.0,<44.0

# Static files
whitenoise>=6.7,<7.0

# API
djangorestframework>=3.15,<4.0

# Dev/Test
ruff>=0.6,<1.0
mypy>=1.11,<2.0
django-stubs>=5.0,<6.0
pytest>=8.3,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0

# Monitoring
sentry-sdk[django]>=2.13,<3.0
